I refactored the tags page to fix prerender and topLevelAwait errors.
- I converted `app/tags/page.tsx` to a Server Component to handle data fetching (`getAllTodos`, `getAllTags`).
- I created a new Client Component `app/tags/tag-page-client.tsx` to manage UI, state (`tagList`), and interactions (`handleCreateTag`).
- The Server Component now fetches data and passes it as props to the Client Component.
- This resolves the 'topLevelAwait' error and the error 'Server Functions cannot be called during initial render' by adhering to Next.js patterns for Server and Client Components.
